+++ src/combined/ascon-aead-masked.c
@@ -52,7 +52,7 @@
  */
 static void ascon_absorb_masked_8
     (ascon_state_t *state, const unsigned char *data,
-     unsigned long long len, uint8_t first_round)
+     size_t len, uint8_t first_round)
 {
 #if ASCON_SLICED
     unsigned char padded[8];
@@ -92,7 +92,7 @@ static void ascon_absorb_masked_8
  */
 static void ascon_absorb_masked_16
     (ascon_state_t *state, const unsigned char *data,
-     unsigned long long len, uint8_t first_round)
+     size_t len, uint8_t first_round)
 {
 #if ASCON_SLICED
     unsigned char padded[16];
@@ -135,7 +135,7 @@ static void ascon_absorb_masked_16
  */
 static void ascon_encrypt_masked_8
     (ascon_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len, uint8_t first_round)
+     const unsigned char *src, size_t len, uint8_t first_round)
 {
 #if ASCON_SLICED
     unsigned char padded[8];
@@ -177,7 +177,7 @@ static void ascon_encrypt_masked_8
  */
 static void ascon_encrypt_masked_16
     (ascon_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len, uint8_t first_round)
+     const unsigned char *src, size_t len, uint8_t first_round)
 {
 #if ASCON_SLICED
     unsigned char padded[16];
@@ -221,7 +221,7 @@ static void ascon_encrypt_masked_16
  */
 static void ascon_decrypt_masked_8
     (ascon_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len, uint8_t first_round)
+     const unsigned char *src, size_t len, uint8_t first_round)
 {
 #if ASCON_SLICED
     unsigned char padded[8];
@@ -263,7 +263,7 @@ static void ascon_decrypt_masked_8
  */
 static void ascon_decrypt_masked_16
     (ascon_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len, uint8_t first_round)
+     const unsigned char *src, size_t len, uint8_t first_round)
 {
 #if ASCON_SLICED
     unsigned char padded[16];
@@ -304,9 +304,9 @@ static void ascon_decrypt_masked_16
 #endif
 
 int ascon128_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k)
@@ -362,10 +362,10 @@ int ascon128_masked_aead_encrypt
 }
 
 int ascon128_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k)
 {
@@ -423,9 +423,9 @@ int ascon128_masked_aead_decrypt
 }
 
 int ascon128a_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k)
@@ -481,10 +481,10 @@ int ascon128a_masked_aead_encrypt
 }
 
 int ascon128a_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k)
 {
@@ -542,9 +542,9 @@ int ascon128a_masked_aead_decrypt
 }
 
 int ascon80pq_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k)
@@ -602,10 +602,10 @@ int ascon80pq_masked_aead_encrypt
 }
 
 int ascon80pq_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k)
 {
@@ -676,7 +676,7 @@ int ascon80pq_masked_aead_decrypt
  */
 static void ascon_absorb_masked_8
     (ascon_masked_state_t *state, const unsigned char *data,
-     unsigned long long len, uint8_t first_round)
+     size_t len, uint8_t first_round)
 {
     unsigned char padded[8];
     unsigned temp;
@@ -704,7 +704,7 @@ static void ascon_absorb_masked_8
  */
 static void ascon_absorb_masked_16
     (ascon_masked_state_t *state, const unsigned char *data,
-     unsigned long long len, uint8_t first_round)
+     size_t len, uint8_t first_round)
 {
     unsigned char padded[16];
     unsigned temp;
@@ -735,7 +735,7 @@ static void ascon_absorb_masked_16
  */
 static void ascon_encrypt_masked_8
     (ascon_masked_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len,
+     const unsigned char *src, size_t len,
      uint8_t first_round)
 {
     unsigned char padded[8];
@@ -768,7 +768,7 @@ static void ascon_encrypt_masked_8
  */
 static void ascon_encrypt_masked_16
     (ascon_masked_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len,
+     const unsigned char *src, size_t len,
      uint8_t first_round)
 {
     unsigned char padded[16];
@@ -805,7 +805,7 @@ static void ascon_encrypt_masked_16
  */
 static void ascon_decrypt_masked_8
     (ascon_masked_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len,
+     const unsigned char *src, size_t len,
      uint8_t first_round)
 {
     unsigned char padded[8];
@@ -839,7 +839,7 @@ static void ascon_decrypt_masked_8
  */
 static void ascon_decrypt_masked_16
     (ascon_masked_state_t *state, unsigned char *dest,
-     const unsigned char *src, unsigned long long len,
+     const unsigned char *src, size_t len,
      uint8_t first_round)
 {
     unsigned char padded[16];
@@ -868,9 +868,9 @@ static void ascon_decrypt_masked_16
 }
 
 int ascon128_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k)
@@ -915,10 +915,10 @@ int ascon128_masked_aead_encrypt
 }
 
 int ascon128_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k)
 {
@@ -965,9 +965,9 @@ int ascon128_masked_aead_decrypt
 }
 
 int ascon128a_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k)
@@ -1012,10 +1012,10 @@ int ascon128a_masked_aead_encrypt
 }
 
 int ascon128a_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k)
 {
@@ -1062,9 +1062,9 @@ int ascon128a_masked_aead_decrypt
 }
 
 int ascon80pq_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k)
@@ -1111,10 +1111,10 @@ int ascon80pq_masked_aead_encrypt
 }
 
 int ascon80pq_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k)
 {
